+++ x/observer/client/cli/query_ballot.go
@@ -1,6 +1,9 @@
 package cli
 
 import (
+	"context"
+	"strconv"
+
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/spf13/cobra"
@@ -73,5 +76,40 @@ func CmdAllBallots() *cobra.Command {
 	}
 	flags.AddQueryFlagsToCmd(cmd)
 	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
+	return cmd
+}
+
+func CmdBallotListForHeight() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "list-ballots-for-height [height]",
+		Short: "Query BallotListForHeight",
+		Args:  cobra.ExactArgs(1),
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+			clientCtx, err := client.GetClientQueryContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			queryClient := types.NewQueryClient(clientCtx)
+
+			height, err := strconv.ParseInt(args[0], 10, 64)
+			if err != nil {
+				return err
+			}
+
+			params := &types.QueryBallotListForHeightRequest{
+				Height: height,
+			}
+
+			res, err := queryClient.BallotListForHeight(context.Background(), params)
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
 	return cmd
 }